Fixtures can now optionally get the test context.
If a fixture is created with a "context" parameter, it will be passed
the test context.  This can be used as a way to pass parameters in to
the fixture.  Note that the fixture can't affect the contents of the
context, and the context will not contain any of the other fixtures at
this point.