First stage of `ExUnitFixtures.AutoLoad`.
This commit adds the beginnings of `ExUnitFixtures.AutoLoad` - a module
that can be used to automatically load fixture named `fixtures.exs` in
the current and parent directories.

This initial implementation is intended to provide a relatively easy
way to load `FixtureModule` code, without having to resort to manually
calling `Code.load_file` on each of the files that define your
FixtureModules.

I'll probably build on this in a future commit to add (optional?)
functionality that automatically imports any fixture modules that are
loaded as part of this, but that is still to be decided upon.  The
docstring is currently written as if I've already implemented this.  I
will need to edit it (and other docs) if I end up implementing it
differently.