Vastly simplified ExUnitFixtures.AutoLoad.
This removes half of the code in the old AutoLoad module, in favour of a
simpler idea.  The AutoLoad module is now used just for loading fixture
files underneath the top directory.

This is better than the old scheme of traversing up the directory
heirarchy, as we only need to traverse things once and we can be sure
we've got all the fixture files.  It also requires a lot less code.