Add ModuleStore that maps fixture modules to files.
This adds a persistent "ModuleStore" that can be used to lookup details
about all registered fixtures.  When used in combination with the
AutoLoad functionality this should make it relatively easy to
automatically use fixtures based on a files "scope".

FixtureModules will automatically register themselves with the
ModuleStore when they are declared.  For now this means that users have
to start the ExUnitFixtures application themselves in their
test_helper.exs file, though we could consider doing this automatically
when declaring a fixture module.