Add ExUnitFixtures.AutoImport
This adds ExUnitFixtures.AutoImport.  When you use this module, it will
look through all parent directories of the current directory until it
finds the mix root, and automatically import any fixture modules that it
finds that are named `fixtures.exs`.

This can be used to build py.test conftest style hierarchies of
fixtures fairly easily.  The code still needs tidied up a bit, and I
might rework _precisely_ how it works, but it's a decent first pass.