Raise a useful error when forget to start server.
Forgot to add some calls to check_server_running in the module store
implementation leading to some not so useful errors in certain
circumstances.

Did attempt to add some Application.ensure_all_started calls in fixture
modules but this didn't work out - ExUnitFixtures loads fixture modules
on startup so this lead to an annoying deadlock.

Fixes #16