Correct default scope to :test not :function.
The documentation mostly states that the default scope for fixtures is
`:test` scoped, but most of the code that actually sets the value was
setting it to `:function` scope.

This commit fixes that.

This was causing an issue with the dependency validation code that
attempted to stop module scoped fixtures from being able to depend on
test scoped fixtures, as fixtures were never actually test scoped, but
function scoped.

Fixes #28.

Also added some tests of compile failures, since that's the only way to
test this.