Module fixtures now only created once (Fixes #30)
Prior to this change, the `create_fixtures` function wasn't making any
decisions about which fixtures it needed to actually create, and was
just assuming that it should create all the fixtures in the dep chain.

This was incorrect - when creating test fixtures, module fixtures should
have already been created, so should not need to be created again.

This fixes that issue, and adds a test to make sure that it doesn't
happen again.