Add a simple ecto sample project.
This adds a simple sample project that uses ecto.  It's fairly bare
bones right now, and borrows it's schema definitions from the ecto
sample project.

It doesn't do a ton right now, but I'm hoping to work on it a bit more
as I build other samples.

It's also helping to demonstrate which features it would be
useful/essential to have in the library, as well as finding some bugs.
A good exercise basically.