Refactor fixture creation for greater flexibility (Fixes #31)
This commit contains a fairly comprehensive refactor of the way fixtures
are created - all fixtures are now created in the tests setup function,
regardless of how they are scoped.

Module scoped fixtures will be stored by a FixtureStore process, which
will be started for each test, and shutdown again afterwards.

This should mean that module scoped fixtures are no longer implicitly
autouse, because they will only be created when they are explicitly
required by a test.  They will still persist throughout the rest of the
run of that file, but don't think there's much I can do about that...