Add Teardown module.
Now that all module fixtures are created as part of the setup for a
specific test, we can't use the on_exit function to tear them down.
Doing so would cause them to be destructed as soon as the test that
created them was stopped.

This commit adds `ExUnitFixtures.teardown` function, which will be
scoped automatically as part of the tests.  This does the same job, but
handles module scoped fixtures as well.