Add `register_fixture` macro.
This starts the API change work described in #15.  The initial step is
just to provide a `register_fixture` macro that decouples the fixture
function creation from the fixture registration.  This is a small
change, and leaves most of the actual implementation alone.

However this small change enables a few things:

- Pattern matching & guards on arguments for fixtures.  (#9)
- Decoupling the name of the arguments used in the function with the
  actual fixture names.  This is particularly useful when fixtures are
  required for side effects and aren't actually used - you can just `_`
  out the argument and the compiler won't complain. (#24)

At the moment, I've chosen not to remove the `deffixture` macro.  I was
considering it, but for now I'm keeping it around for the convenient
simpler cases.