AutoImport should not import non-directories.
If `use ExUnitFixtures.AutoImport` ends up being called from a doctest
in elixir 1.3.0 the path will have `(doctest for)` or similar on the end
of it.  Obviously this is not a directory, so when we call `File.ls!` on
it, things fail.  This checks that the path is actually a directory
before calling `File.ls!`